requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.1.4
django-bootstrap-v5==1.0.11
gunicorn==20.1.0
Pillow==9.3.0
python-dateutil==2.8.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
